"proper" error management
I clearly was not understanding how appendSlice yours and ignoring
completely a few copies. Allocating strings and freeing them with this
approach is rather simple, no need to depend on an arena allocator or
anything. Each string that gets allocated can be freed after being
copied and then it's simply a matter of freeing the final string that
was returned. Quite "clean" actually.